fix: move save layout out of close event
